+++ spec/support/database.rb
@@ -1,7 +1,11 @@
 require 'active_record'
 
-ActiveRecord::Base.establish_connection(
+connection_config = {
   adapter: 'postgresql',
-  database: 'real_data_tests_test',
-  host: 'localhost'
-)
+  database: ENV.fetch('PGDATABASE', 'real_data_tests_test')
+}
+connection_config[:host] = ENV['PGHOST'] if ENV['PGHOST']
+connection_config[:username] = ENV['PGUSER'] if ENV['PGUSER']
+connection_config[:password] = ENV['PGPASSWORD'] if ENV['PGPASSWORD']
+
+ActiveRecord::Base.establish_connection(connection_config)